+++ apps/site/src/lib/docs-pages/aliases.svelte
@@ -1,0 +1,21 @@
+<script lang="ts">
+    import { base } from "$app/paths";
+</script>
+
+<h1>Aliases</h1>
+
+<p>
+    <strong>Aliases</strong> are named pre-defined actions.
+</p>
+<p>
+    An example of a useful alias might be a custom vibration pattern that you
+    want to reuse paired with certain kinds of actions in a macro. An alias will
+    allow you to define a name for a sequence of actions and reuse it throughout
+    your macros rather than redefining it each time.
+</p>
+
+<img
+    class="m-2 shadow-2xl shadow-gray-700 rounded-2xl"
+    src="{base}/images/aliases.png"
+    alt="Aliases"
+/>