+++ apps/site/src/lib/docs-pages/autoswitch.svelte
@@ -1,0 +1,17 @@
+<script lang="ts">
+    import { base } from "$app/paths";
+</script>
+
+<h1>Auto-Switch</h1>
+
+<p>
+    <strong>Auto-Switch</strong> defines rules for automatically switching between
+    profiles when a specific application is active or gains focus. For example you
+    can set it up so that when you open a web browser, MapXr automatically switches
+    to a profile optimized for browsing.
+</p>
+<img
+    class="m-2 shadow-2xl shadow-gray-700 rounded-2xl"
+    src="{base}/images/autoswitch.png"
+    alt="Auto-Switch"
+/>