+++ apps/site/src/lib/docs-pages/debug.svelte
@@ -1,0 +1,14 @@
+<script lang="ts">
+    import { base } from "$app/paths";
+</script>
+
+<h1>Debug</h1>
+
+<p>
+    <strong>Debug</strong> useful when troubleshooting what might be going wrong because
+    nothing is perfect all the time. Simply turn on the debug mode on the top of the
+    screen and MapXr will display what tap it detected and what if any matches it
+    found.
+</p>
+
+<img class="m-2" src="{base}/images/debug.png" alt="Debug" />