+++ apps/site/src/lib/docs-pages/profilesettings.svelte
@@ -1,0 +1,18 @@
+<script lang="ts">
+    import { base } from "$app/paths";
+</script>
+
+<h1>Profile Settings</h1>
+
+<p>
+    Profile <strong>Settings</strong> allows you to adjust timings of multi hand combinations
+    and double/triple tap detection windows as well as the window in which tap sequences
+    are recognized.
+</p>
+
+<p>Additionally, this is where you can toggle layer passthrough</p>
+<img
+    class="m-2 shadow-2xl shadow-gray-700 rounded-2xl"
+    src="{base}/images/profilesettings.png"
+    alt="Profile Settings"
+/>