+++ apps/site/src/lib/docs-pages/settings.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+    import { base } from "$app/paths";
+</script>
+
+<h1>Settings</h1>
+
+<p>
+    <strong>Settings</strong> allows you to adjust the behavior of MapXr inlcuding
+    starting minimized, notifications and vibration settings.
+</p>
+<img
+    class="m-2 shadow-2xl shadow-gray-700 rounded-2xl"
+    src="{base}/images/settings.png"
+    alt="Settings"
+/>