+++ apps/site/src/lib/docs-pages/variables.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+    import { base } from "$app/paths";
+</script>
+
+<h1>Variables</h1>
+
+<p>
+    <strong>Variables</strong> are named values that hold a True|False value that
+    can be used to have conditional behavior in your mappings.
+</p>
+<img
+    class="m-2 shadow-2xl shadow-gray-700 rounded-2xl"
+    src="{base}/images/settings.png"
+    alt="Settings"
+/>